packages/agents/pyproject.toml
[project]
name = "ai-base-agent"
version = "0.1.0"
description = "Base agent class for AI Infrastructure"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ai-core",
    "ai-messaging",
    "ai-memory",
    "pydantic>=2.5.0",
    "aider-chat>=0.70.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/base_agent"]
